+++ src/main/java/com/mredust/oj/model/dto/problem/ProblemUpdateRequest.java
@@ -67,17 +67,17 @@ public class ProblemUpdateRequest implements Serializable {
     /**
      * 运行时间限制（ms）
      */
-    private Integer runTime;
+    private Long runTime;
     
     /**
      * 内存限制（KB）
      */
-    private Integer runMemory;
+    private Long runMemory;
     
     /**
      * 栈大小（KB）
      */
-    private Integer runStack;
+    private Long runStack;
     
     
     /**
